refactor: 코리 반영
