+++ src/History/components/historyData/HistoryList.tsx
@@ -2,7 +2,9 @@ import { css } from '@emotion/react';
 import { useState } from 'react';
 
 import { ObjectiveTypes } from '../../type/historyData';
-import { HistoryKeyResult, HistoryObjective, HistoryTask } from './ItemWrappers';
+import HistoryKeyResult from './historyItemWrappers/HistoryKeyResult';
+import HistoryObjective from './historyItemWrappers/HistoryObjective';
+import { HistoryTask } from './historyItemWrappers/HistoryTask';
 
 const HistoryList = ({ krList, objCategory, objPeriod, objProgress, objTitle }: ObjectiveTypes) => {
   const [isShowKR, setIsShowKR] = useState(false);
